requirements.txt
numpy==1.26.4
python-dotenv==1.0.1
torch==2.3.0
# torchinfo==1.8.0
torchvision==0.18.0
tqdm==4.66.4
typing_extensions==4.9.0
scipy==1.13.0
pillow==10.3.0
